+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file, following t
 - reorganised everything to use es6 modules
 - read bondOrder from 'pdbx_value_order' in mmcif files
 - interpret 'CONECT 1529 1528 1528' as double bond in pdb files
+- `side` Buffer/Representation parameter changed: THREE.FrontSide => "front", THREE.BackSide => "back", THREE.DoubleSide => "double"
 
 ### Removed
 - python-based build scripts